backend/requirements.txt
# Flask and web framework
Flask==2.2.3
Flask-Cors==3.0.10
Werkzeug==2.2.3
itsdangerous==2.2.0
Jinja2==3.1.6

# Database
pg8000==1.29.4
psycopg2-binary==2.9.10

# Authentication
PyJWT==2.6.0
jwt==1.3.1
Flask-JWT-Extended==4.5.3

# Environment and configuration
python-dotenv==0.21.0

# HTTP and API
requests==2.28.2

# Security
cryptography==38.0.4

# PDF generation
reportlab==4.0.9

# Excel generation
openpyxl==3.1.2
XlsxWriter==3.2.3

# Date handling
python-dateutil==2.8.2
pytz==2025.2

# Data manipulation
pandas==2.2.3
numpy==2.2.5
